LPS-55532 Fix unit tests.
