+++ portal-impl/test/unit/com/liferay/portal/template/ClassLoaderResourceParserTest.java
@@ -81,39 +81,4 @@ public class ClassLoaderResourceParserTest {
 		}
 	}
 
-	@Test
-	public void testNormalizePath() {
-		Assert.assertEquals(
-			"abc", ClassLoaderResourceParser.normalizePath("abc"));
-		Assert.assertEquals(
-			"/abc", ClassLoaderResourceParser.normalizePath("/abc"));
-
-		try {
-			ClassLoaderResourceParser.normalizePath("//");
-
-			Assert.fail();
-		}
-		catch (IllegalArgumentException iae) {
-			Assert.assertEquals("Unable to parse path //", iae.getMessage());
-		}
-
-		Assert.assertEquals(
-			"abc", ClassLoaderResourceParser.normalizePath("abc/./"));
-		Assert.assertEquals(
-			"def", ClassLoaderResourceParser.normalizePath("abc/../def"));
-
-		try {
-			ClassLoaderResourceParser.normalizePath("../");
-
-			Assert.fail();
-		}
-		catch (IllegalArgumentException iae) {
-			Assert.assertEquals("Unable to parse path ../", iae.getMessage());
-		}
-
-		Assert.assertEquals(
-			"/efg/hij",
-			ClassLoaderResourceParser.normalizePath("/abc/../efg/./hij/"));
-	}
-
 }
